Task #4959: Reintegrated task branch (DelayAndBandPassKernel: wrong subbandFreq type passed) with the trunk

+++ trunk/RTCP/Cobalt/GPUProc/src/cuda/Kernels/DelayAndBandPassKernel.cc
@@ -86,7 +86,7 @@ namespace LOFAR
     }
 
 
-    void DelayAndBandPassKernel::enqueue(PerformanceCounter &counter, float subbandFrequency, size_t SAP)
+    void DelayAndBandPassKernel::enqueue(PerformanceCounter &counter, double subbandFrequency, size_t SAP)
     {
       setArg(2, subbandFrequency);
       setArg(3, SAP);
